Rename `LRBv21` -> `solver`
